# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKF Racing Hub API",
    "description": "REST API of the SKF Racing Hub site.\n\n**Authentication.** A `session_id` cookie, set by the Discord login. Operations\nthat need a session list it under *security*; the roles they accept are in\n`x-required-roles`.\n\n**Errors.** Every error is an `application/problem+json` document (`Problem`).\n`code` is a stable machine-readable identifier, `detail` a message that can be\nshown to the user. Validation failures (422) also carry `errors`, one entry per\nrejected field.\n\n**Paging.** Collections that take `limit` and `offset` return a plain array;\nthe total is in `X-Total-Count` and neighbouring pages in an RFC 8288 `Link`\nheader.\n\n**Tracing.** Every response carries `X-Request-ID`; quote it when reporting a\nproblem.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/championships": {
      "get": {
        "tags": [
          "Championships"
        ],
        "summary": "List championships",
        "description": "The championships shown on the site, in SimGrid's order.",
        "operationId": "listChampionships",
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "inactive",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`inactive` adds the championships hidden from the site; admins only."
            },
            "description": "`inactive` adds the championships hidden from the site; admins only."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionshipSummaryOut"
                  }
                }
              }
            },
            "headers": {
              "X-Data-Stale": {
                "description": "`true` when SimGrid was unreachable and cached data was served.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/api/v1/championships/{championshipId}": {
      "patch": {
        "tags": [
          "Championships"
        ],
        "summary": "Update championship",
        "description": "Show or hide the championship on the site. Everything else about it lives in SimGrid.\n\nRequires role: `admin`, `super_admin`.",
        "operationId": "updateChampionship",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChampionshipUpdate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ]
      },
      "get": {
        "tags": [
          "Championships"
        ],
        "summary": "Get championship",
        "operationId": "getChampionship",
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionshipOut"
                }
              }
            },
            "headers": {
              "X-Data-Stale": {
                "description": "`true` when SimGrid was unreachable and cached data was served.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
...
    "/api/v1/championships/{championshipId}/races": {
      "get": {
        "tags": [
          "Championships"
        ],
        "summary": "List championship races",
        "description": "All races of the championship, including future ones.",
        "operationId": "listChampionshipRaces",
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionshipRaceOut"
                  }
                }
              }
            },
            "headers": {
              "X-Data-Stale": {
                "description": "`true` when SimGrid was unreachable and cached data was served.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
...
    "/api/v1/custom-championships": {
      "get": {
        "tags": [
          "Custom championships"
        ],
        "summary": "List custom championships",
        "operationId": "listCustomChampionships",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "communityId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomChampionshipOut"
                  }
                }
              }
            },
            "headers": {
              "X-Total-Count": {
                "description": "Size of the whole collection, across all pages.",
                "schema": {
                  "type": "integer"
                }
              },
              "Link": {
                "description": "RFC 8288 links to the `next` and `prev` pages, when they exist.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "post": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Create custom championship",
        "operationId": "createCustomChampionship",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomChampionshipCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomChampionshipOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      }
    },
    "/api/v1/custom-championships/{championshipId}": {
      "get": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Get custom championship",
        "operationId": "getCustomChampionship",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomChampionshipOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "patch": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Update custom championship",
        "operationId": "updateCustomChampionship",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomChampionshipUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomChampionshipOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "delete": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Delete custom championship",
        "operationId": "deleteCustomChampionship",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      }
    },
    "/api/v1/custom-championships/{championshipId}/races": {
      "post": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Create custom race",
        "operationId": "createCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRaceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRaceOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "put": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Replace custom races",
        "description": "Replace the full race list in one request.\n\nRequires role: `admin`, `super_admin`, `community_manager`.",
        "operationId": "replaceCustomRaces",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustomRaceUpsert"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomRaceOut"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ]
      }
    },
    "/api/v1/custom-championships/{championshipId}/races/{raceId}": {
      "get": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Get custom race",
        "operationId": "getCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRaceOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "patch": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Update custom race",
        "operationId": "updateCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRaceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRaceOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "delete": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Delete custom race",
        "operationId": "deleteCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionshipOut": {
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "spotsTaken": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "acceptingRegistrations": {
            "type": "boolean",
            "default": false
          },
          "hostName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "gameName": {
            "type": "string",
            "default": ""
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "resultsUrl": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "discordUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "roundNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "allRoundsNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "image",
          "startDate",
          "endDate",
          "capacity",
          "spotsTaken",
          "acceptingRegistrations",
          "hostName",
          "gameName",
          "url",
          "resultsUrl",
          "discordUrl",
          "roundNumber",
          "allRoundsNumber"
        ]
      },
...
      "ChampionshipSummaryOut": {
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "acceptingRegistrations": {
            "type": "boolean",
            "default": false
          },
          "eventCompleted": {
            "type": "boolean",
            "description": "SimGrid marks the championship as finished.",
            "default": false
          },
          "isActive": {
            "type": "boolean",
            "description": "Shown on the site; inactive ones are admin-only.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "acceptingRegistrations",
          "eventCompleted",
          "isActive"
        ]
      },
      "ChampionshipUpdate": {
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "type": "object",
        "required": [
          "isActive"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session id, set as a cookie by the Discord login.",
        "in": "cookie",
        "name": "session_id"
      },
      "ingestToken": {
        "type": "http",
        "description": "Shared token of the external incident parsers.",
        "scheme": "bearer"
      }
    },
    "responses": {
      "BadGateway": {
        "description": "Bad Gateway",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "UnprocessableContent": {
        "description": "Unprocessable Content",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblem"
            }
          }
        }
      }
    }
  },
...
}